+++ dart/lib/src/scope.dart
@@ -44,16 +44,6 @@ class Scope {
   /// Get the current user.
   SentryUser? get user => _user;
 
-  /// Sets the current user
-  /// This method is deprecated, use the [setUser(user)] instead.
-  /// This method will be removed in the future.
-  /// The breaking change is due to the [enableScopeSync] feature that
-  /// requires returning a [Future].
-  @Deprecated('Use [setUser(user)] instead')
-  set user(SentryUser? user) {
-    _user = user;
-  }
-
   void _setUserSync(SentryUser? user) {
     _user = user;
   }
@@ -150,9 +140,6 @@ class Scope {
   final List<SentryAttachment> _attachments = [];
 
   List<SentryAttachment> get attachments => List.unmodifiable(_attachments);
-
-  @Deprecated('Use [attachments] instead')
-  List<SentryAttachment> get attachements => attachments;
 
   Scope(this._options);